+++ src/EngineCore/Events/Details/TimeEventDetails.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Sovereign Engine
- * Copyright (c) 2018 opticfluorine
+ * Copyright (c) 2019 opticfluorine
  *
  * Permission is hereby granted, free of charge, to any person obtaining a 
  * copy of this software and associated documentation files (the "Software"), 
@@ -21,42 +21,19 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-namespace Sovereign.EngineCore.Util
+namespace Sovereign.EngineCore.Events.Details
 {
 
     /// <summary>
-    /// Constants for unit conversions.
+    /// Event details recording a system time.
     /// </summary>
-    public static class Units
+    public sealed class TimeEventDetails : IEventDetails
     {
-
+        
         /// <summary>
-        /// Unit conversion constants related to system time.
+        /// System time.
         /// </summary>
-        public static class SystemTime
-        {
-
-            /// <summary>
-            /// One minute in system time.
-            /// </summary>
-            public const ulong Minute = 60 * Second;
-
-            /// <summary>
-            /// One second in system time.
-            /// </summary>
-            public const ulong Second = 1000 * Millisecond;
-
-            /// <summary>
-            /// One ms in system time.
-            /// </summary>
-            public const ulong Millisecond = 1000 * Microsecond;
-
-            /// <summary>
-            /// One us in system time.
-            /// </summary>
-            public const ulong Microsecond = 1;
-
-        }
+        public ulong SystemTime { get; set; }
 
     }