+++ src/Performance/Performance/IPerformanceMonitor.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Sovereign Engine
- * Copyright (c) 2018 opticfluorine
+ * Copyright (c) 2019 opticfluorine
  *
  * Permission is hereby granted, free of charge, to any person obtaining a 
  * copy of this software and associated documentation files (the "Software"), 
@@ -21,42 +21,27 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-namespace Sovereign.EngineCore.Util
+using Sovereign.EngineCore.Events;
+
+namespace Sovereign.Performance.Performance
 {
 
     /// <summary>
-    /// Constants for unit conversions.
+    /// Interface for classes that monitor engine performance.
     /// </summary>
-    public static class Units
+    public interface IPerformanceMonitor
     {
 
         /// <summary>
-        /// Unit conversion constants related to system time.
+        /// Called when a performance event occurs.
         /// </summary>
-        public static class SystemTime
-        {
+        /// <param name="eventDetails"></param>
+        void OnPerformanceEvent(IEventDetails eventDetails);
 
-            /// <summary>
-            /// One minute in system time.
-            /// </summary>
-            public const ulong Minute = 60 * Second;
-
-            /// <summary>
-            /// One second in system time.
-            /// </summary>
-            public const ulong Second = 1000 * Millisecond;
-
-            /// <summary>
-            /// One ms in system time.
-            /// </summary>
-            public const ulong Millisecond = 1000 * Microsecond;
-
-            /// <summary>
-            /// One us in system time.
-            /// </summary>
-            public const ulong Microsecond = 1;
-
-        }
+        /// <summary>
+        /// Called when one minute of system time has elapsed.
+        /// </summary>
+        void ReportPerformance();
 
     }